Add helm chart (#97)
* Add Helm chart for Kubernetes and update Dockerfile command

This commit adds a Helm chart to the project structure to support management on Kubernetes. It also modifies the Dockerfile's CMD to use the "discover" command. The Helm chart includes various configuration files for deployment, services, ingress, and tests. It also introduces a new ConfigMap and updates the .codecov.yml file to ignore the helm-chart directory.

* Update CMD instruction in Dockerfile

The CMD instruction in the Dockerfile was modified by removing the "discover" parameter. This alteration changes the default command that will be run when the Docker container is started.